pyproject.toml
[tool.poetry]
name = "dota-timings-hud"
version = "0.1.0"
description = "Small always-on-top HUD for Dota 2 timings (manual start)"
authors = ["You"]
readme = "README.md"
packages = [{ include = "dota_hud", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10"
keyboard = "^0.13.5"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
dota-hud = "dota_hud.__main__:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.5"
mypy = "^1.13.0"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
